Updated the logic of the form to allow reporting without UPRN.
User is forced to fill in the location description in this scenario and select the propery type themselves.

You are still presented with the list of addresses based on the postcode of the location you selected and if you select an address it defaults back to the standard with address flow.

This is a WIP commit as we may need to redesign.
Main issues are:
 - Should we be using the Address.razor page or spliting to seperate locations?
 - Do we need to include a map image above the location box

Tested flow from beginning to end without saving in all contexts and believe you can get to summary page.

+++ Database/Extensions/EligibilityCheckExtensions.cs
@@ -6,11 +6,12 @@ public static class EligibilityCheckExtensions
 {
     internal static EligibilityCheckCreated ToMessageCreated(this EligibilityCheck eligibilityCheck, string reference)
     {
+        //TODO - sort UPRN
         return new EligibilityCheckCreated(
             eligibilityCheck.Id,
             reference,
             eligibilityCheck.CreatedUtc,
-            eligibilityCheck.Uprn,
+            (long)eligibilityCheck.Uprn,
             eligibilityCheck.Easting,
             eligibilityCheck.Northing,
             eligibilityCheck.ImpactStart,
@@ -23,10 +24,11 @@ public static class EligibilityCheckExtensions
 
     internal static EligibilityCheckUpdated ToMessageUpdated(this EligibilityCheck eligibilityCheck)
     {
+        //TODO - sort UPRN
         return new EligibilityCheckUpdated(
             eligibilityCheck.Id,
             eligibilityCheck.UpdatedUtc ?? DateTimeOffset.UtcNow,
-            eligibilityCheck.Uprn,
+            (long)eligibilityCheck.Uprn,
             eligibilityCheck.Easting,
             eligibilityCheck.Northing,
             eligibilityCheck.ImpactStart,
